backend/requirements-solana.txt
solana==0.30.2
solders==0.18.1
spl-token==0.4.0
